Move WarnAboutIsolateLockScope to thread-scopes.h
